dom/base/test/meta_viewport/test_meta_viewport_empty_content_and_valid_content_tags.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>a valid meta viewport tag and empty content attribute viewport tag</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="width=400">
  <meta name="viewport" content="">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <script type="application/javascript">
    "use strict";

    add_task(async function multiple_viewport_tags() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      // The empty content attribute meta tag should be valid, which means
      // it supersedes the first `width=400`.
      is(info.width, 980, "width should be default, 980");
    });
  </script>
</body>
</html>
